v2: applying patch from FObermaier, many more tests pass, many thanks Felix!
git-svn-id: http://nettopologysuite.googlecode.com/svn/branches/v2.0@458 53ca42c3-9f19-0410-8ccb-b990a9bb5db6

+++ NetTopologySuite/GeometriesGraph/Index/SweepLineEvent.cs
@@ -88,7 +88,7 @@ namespace GisSharpBlog.NetTopologySuite.GeometriesGraph.Index
                 return 1;
             }
 
-            return _eventType.CompareTo(other._eventType);
+            return ((Int32)_eventType).CompareTo((Int32)other._eventType);
         }
 
         #endregion
